show notifications

+++ core/Friends.class.php
@@ -27,7 +27,6 @@ class Friends {
         $friend_req = mysqli_fetch_assoc($result);
         if ($friend_req) {
             // Accept
-            $this->_db->query("DELETE FROM friend_request WHERE (user_id={$user_id} AND user_friend_id={$friend_id}) OR (user_id={$friend_id} AND user_friend_id={$user_id})");
             $this->_db->query("INSERT INTO friends (user_id, user_friend_id, friendship_created) VALUES ({$user_id}, {$friend_id}, CURRENT_TIME())");
             $this->_db->query("INSERT INTO friends (user_id, user_friend_id, friendship_created) VALUES ({$friend_id}, {$user_id}, CURRENT_TIME())");
         } else {
@@ -44,7 +43,7 @@ class Friends {
      * @return bool
      */
     public function deleteFriend($user_id, $friend_id) {
-        $this->_db->query("DELETE FROM friend_request WHERE user_id={$user_id} AND user_friend_id={$friend_id}");
+        $this->_db->query("DELETE FROM friend_request WHERE (user_id={$user_id} AND user_friend_id={$friend_id}) OR (user_id={$friend_id} AND user_friend_id={$user_id})");
         $this->_db->query("DELETE FROM friends WHERE (user_id={$user_id} AND user_friend_id={$friend_id} OR user_id={$friend_id} AND user_friend_id={$user_id})");
         return true;
     }
@@ -98,7 +97,7 @@ class Friends {
         $result = $this->_db->query("SELECT 1 FROM friend_request WHERE user_id={$user_id} AND user_friend_id={$friend_id} LIMIT 1");
         $friend_req = mysqli_fetch_assoc($result);
         $status['is_friend'] = $is_friend;
-        $status['is_friend_req_sent'] = $friend_req;
+        $status['is_friend_req_sent'] = !$is_friend && $friend_req;
         return $status;
     }
 
@@ -108,12 +107,57 @@ class Friends {
      * @return array
      */
     public function getFriendRequests($user_id) {
-        $query = "SELECT * FROM friend_request INNER JOIN users_info ON friend_request.user_id=users_info.user_id WHERE friend_request.user_friend_id={$user_id}";
+        $query = "
+          SELECT * FROM friend_request
+          LEFT JOIN friends ON friend_request.user_friend_id=friends.user_friend_id
+          INNER JOIN users_info ON friend_request.user_id=users_info.user_id
+          WHERE friend_request.user_friend_id={$user_id} AND friends.user_id IS NULL";
         $result = $this->_db->query($query);
         $friend_requests = [];
         while ($row = mysqli_fetch_assoc($result)) {
             $friend_requests[] = $row;
         }
         return $friend_requests;
+    }
+
+    /**
+     * Get user's notifications (friend request received, friend added, someone commented or liked on my post)
+     * @param int $offset
+     * @param int $limit
+     * @return array
+     */
+    public function getNotifications($offset = 0, $limit = 10) {
+        $user_id = $_SESSION['auth']['user_id'];
+        $sql_request_sent = "
+            SELECT friend_request.user_id AS friend_id, CONCAT(user_firstname, ' ', user_lastname) AS full_name, user_profile_picture, 'request_sent' AS type, request_created AS date, NULL AS post_id
+            FROM friend_request
+            INNER JOIN users_info ON friend_request.user_id=users_info.user_id
+            WHERE friend_request.user_friend_id={$user_id}";
+        $sql_friend_added = "
+            SELECT friends.user_friend_id AS friend_id, CONCAT(user_firstname, ' ', user_lastname) AS full_name, user_profile_picture, 'friend_added' AS type, friendship_created AS date, NULL AS post_id
+            FROM friends
+            INNER JOIN users_info ON friends.user_friend_id=users_info.user_id
+            WHERE friends.user_id={$user_id}";
+
+        $sql_post_commented = "
+            SELECT comments.user_id AS friend_id, CONCAT(user_firstname, ' ', user_lastname) AS full_name, user_profile_picture, 'post_commented' AS type, comment_time AS date, comments.post_id AS post_id
+            FROM comments
+            INNER JOIN posts ON comments.post_id=posts.post_id
+            INNER JOIN users_info ON comments.user_id=users_info.user_id
+            WHERE posts.user_id={$user_id} AND comments.user_id!={$user_id}";
+
+        $sql_post_liked = "
+            SELECT likes.user_id AS friend_id, CONCAT(user_firstname, ' ', user_lastname) AS full_name, user_profile_picture, 'post_liked' AS type, like_created AS date, likes.post_id AS post_id
+            FROM likes
+            INNER JOIN posts ON likes.post_id=posts.post_id
+            INNER JOIN users_info ON likes.user_id=users_info.user_id
+            WHERE posts.user_id={$user_id} AND likes.user_id!={$user_id}";
+
+        $result = $this->_db->query("SELECT * FROM ({$sql_request_sent} UNION {$sql_friend_added} UNION {$sql_post_commented} UNION {$sql_post_liked}) AS tmp ORDER BY date DESC LIMIT {$offset}, {$limit}");
+        $notifications = [];
+        while ($row = mysqli_fetch_assoc($result)) {
+            $notifications[] = $row;
+        }
+        return $notifications;
     }
 };
